+++ limberframework/database/database_service_provider.py
@@ -1,26 +1,43 @@
 """Database Service Provider
 
 Classes:
-- DatabaseServiceProvider: Manages connections and sessions with databases.
+- DatabaseServiceProvider: Registers database services.
 """
-from typing import Dict
-from sqlalchemy.orm import sessionmaker
+from sqlalchemy.orm import Session, sessionmaker
+from limberframework.database.connections import Connection
 from limberframework.database.connections import make_connection
-from limberframework.support.metaclasses import Singleton
+from limberframework.support.service_providers import ServiceProvider
 
-class DatabaseServiceProvider(metaclass=Singleton):
-    """Manages connections and sessions with databases.
-
-    Attributes:
-    - database Connection -- a Connection object with access to the database engine.
-    - session Session -- a callable Session object to create new sessions.
-    """
-    def __init__(self, config: Dict) -> None:
-        """Establishes a connection to the database and prepares sessions.
-
-        Arguments:
-        - config dict -- connection settings for the database.
+class DatabaseServiceProvider(ServiceProvider):
+    """Registers database services to the service container."""
+    def register(self) -> None:
+        """Registers the database connection and session
+        services to the service container.
         """
-        self.database = make_connection(config)
-        self.session = sessionmaker(autocommit=False, autoflush=False,
-                                    bind=self.database.engine)
+        def register_database_connection(app: 'Application') -> Connection:
+            """Closure for establshing a database connection service.
+
+            Arguments:
+            app Application -- foundation.application.Application object
+
+            Returns:
+            Connection object
+            """
+            return make_connection(app['config']['database'])
+
+        self.app.bind('db.connection', register_database_connection, True)
+
+        def register_database_session(app: 'Application') -> Session:
+            """Closure for establishing a database session
+            using the existing database connection.
+
+            Arguments:
+            app Application -- foundation.application.Application object
+
+            Returns:
+            Session object
+            """
+            return sessionmaker(autocommit=False, autoflush=False,
+                                bind=app['db.connection'].engine)()
+
+        self.app.bind('db.session', register_database_session)