+++ CHANGELOG.md
@@ -10,6 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Database connection and session management for SQLite and PostgreSQL databases.
 - Configuration service provider to manage configuration settings and add dotenv file support.
 - Application class to act as a service container that manages available services.
+- Config class to manage application configuration.
+- DatabaseSessionMiddleware to automatically establish a database session when a request is received.
 
 ### Fixed
 - Resolve SQLite single thread warning by turning off 'same thread check'.
+
+### Changed
+- DatabaseServiceProvider to implement required methods by ServiceProvider for registering services with with the Application class.
+- ConfigServiceProvider to implement required methods by ServiceProvider for registering services with the Application class.
+
+### Removed
+- Singleton metaclass as singletons are now handled by the Application class.